+++ user/views.py
@@ -9,12 +9,14 @@ from airone.lib.http import HttpResponseSeeOther
 from airone.lib.http import http_get, http_post
 from airone.lib.http import render
 from airone.lib.http import check_superuser
+from airone.lib.profile import airone_profile
 
 from rest_framework.authtoken.models import Token
 
 from .models import User
 
 
+@airone_profile
 @http_get
 def index(request):
     if not request.user.is_authenticated():
@@ -31,11 +33,13 @@ def index(request):
     return render(request, 'list_user.html', context)
 
 
+@airone_profile
 @http_get
 def create(request):
     return render(request, 'create_user.html')
 
 
+@airone_profile
 @http_post([
     {'name': 'name', 'type': str, 'checker': lambda x: (
         x['name'] and not User.objects.filter(username=x['name']).exists()
@@ -62,6 +66,7 @@ def do_create(request, recv_data):
     return JsonResponse({})
 
 
+@airone_profile
 @http_get
 def edit(request, user_id):
     current_user = User.objects.get(id=request.user.id)
@@ -84,6 +89,7 @@ def edit(request, user_id):
     return render(request, 'edit_user.html', context)
 
 
+@airone_profile
 @http_post([
     {'name': 'name', 'type': str, 'checker': lambda x: x['name']},
     {'name': 'email', 'type': str, 'checker': lambda x: x['email']},
@@ -130,6 +136,7 @@ def do_edit(request, user_id, recv_data):
     return JsonResponse({})
 
 
+@airone_profile
 @http_get
 def edit_passwd(request, user_id):
 
@@ -152,6 +159,7 @@ def edit_passwd(request, user_id):
     return render(request, 'edit_passwd.html', context)
 
 
+@airone_profile
 @http_post([
     {'name': 'old_passwd', 'type': str, 'checker': lambda x: x['old_passwd']},
     {'name': 'new_passwd', 'type': str, 'checker': lambda x: x['new_passwd']},
@@ -183,6 +191,7 @@ def do_edit_passwd(request, user_id, recv_data):
     return JsonResponse({})
 
 
+@airone_profile
 @http_post([
     {'name': 'new_passwd', 'type': str, 'checker': lambda x: x['new_passwd']},
     {'name': 'chk_passwd', 'type': str, 'checker': lambda x: x['chk_passwd']},
@@ -203,6 +212,7 @@ def do_su_edit_passwd(request, user_id, recv_data):
     return JsonResponse({})
 
 
+@airone_profile
 @http_post([])
 @check_superuser
 def do_delete(request, user_id, recv_data):